requirements.txt
fastapi
uvicorn
pdfplumber
langchain-text-splitters
langgraph
httpx
markdownify
pydantic-settings
PyYAML
